Moved all packets into their own package
All the packets in the top level was getting to be too much
