+++ goqtt.go
@@ -2,12 +2,13 @@ package goqtt
 
 import (
 	"bytes"
+	"github.com/andschneider/goqtt/packets"
 	"log"
 	"net"
 	"time"
 )
 
-func SendPacket(c net.Conn, packet []byte, verbose bool) {
+func sendPacket(c net.Conn, packet []byte, verbose bool) {
 	_, err := c.Write(packet)
 	if verbose {
 		log.Printf("sent packet: %b", packet)
@@ -20,17 +21,17 @@ func SendPacket(c net.Conn, packet []byte, verbose bool) {
 // SendPing is a helper function to create a Ping and send it right away.
 func SendPing(c net.Conn, verbose bool) error {
 	buf := new(bytes.Buffer)
-	ping := CreatePingReqPacket()
+	ping := packets.CreatePingReqPacket()
 	err := ping.Write(buf, verbose)
 	if err != nil {
 		log.Fatal(err)
 		return err
 	}
-	SendPacket(c, buf.Bytes(), verbose)
+	sendPacket(c, buf.Bytes(), verbose)
 
 	// response
 	go func() {
-		_, err = Reader(c)
+		_, err = packets.Reader(c)
 		if err != nil {
 			log.Fatal(err)
 		}
@@ -42,16 +43,16 @@ func SendPing(c net.Conn, verbose bool) error {
 // TODO figure out how to generalize this to other packet types
 func SendConnect(c net.Conn, verbose bool) error {
 	buf := new(bytes.Buffer)
-	cpack := CreateConnectPacket()
+	cpack := packets.CreateConnectPacket()
 	err := cpack.Write(buf, verbose)
 	if err != nil {
 		log.Fatal(err)
 		return err
 	}
-	SendPacket(c, buf.Bytes(), verbose)
+	sendPacket(c, buf.Bytes(), verbose)
 
 	// response
-	_, err = Reader(c)
+	_, err = packets.Reader(c)
 	if err != nil {
 		log.Fatal(err)
 		return err
@@ -62,16 +63,16 @@ func SendConnect(c net.Conn, verbose bool) error {
 // TODO figure out how to generalize this to other packet types
 func SendSubscribe(c net.Conn, topic string, verbose bool) error {
 	buf := new(bytes.Buffer)
-	spack := CreateSubscribePacket(topic)
+	spack := packets.CreateSubscribePacket(topic)
 	err := spack.Write(buf, verbose)
 	if err != nil {
 		log.Fatal(err)
 		return err
 	}
-	SendPacket(c, buf.Bytes(), verbose)
+	sendPacket(c, buf.Bytes(), verbose)
 
 	// response
-	_, err = Reader(c)
+	_, err = packets.Reader(c)
 	if err != nil {
 		log.Fatal(err)
 		return err
@@ -100,7 +101,7 @@ func SubscribeLoop(conn net.Conn) {
 
 	for {
 		//log.Println("start loop")
-		_, err := Reader(conn)
+		_, err := packets.Reader(conn)
 		if err != nil {
 			log.Fatal(err)
 			return
