+++ packets/packet_subscribe_test.go
@@ -1,4 +1,4 @@
-package goqtt
+package packets
 
 import (
 	"fmt"
